testing

+++ platform/src/main/java/ai/enpasos/muzero/platform/agent/e_experience/db/repo/TimestepRepo.java
@@ -117,7 +117,7 @@ public interface TimestepRepo extends JpaRepository<TimeStepDO,Long> {
 
     @Transactional
     @Query(value = "SELECT count(*) FROM  timestep t where box = :box", nativeQuery = true)
-    int numBox(int box);
+    long numBox(int box);
