+++ platform/src/main/java/ai/enpasos/muzero/platform/run/TestRulesLossFromStart.java
@@ -119,7 +119,7 @@ public class TestRulesLossFromStart {
         return true;
     }
 
-    public int numBox( int n) {
+    public long numBox( int n) {
         return timestepRepo.numBox(n);
     }
 }
